Split histogram component
Previously, the `NumericHistogramComponent` queried for multiple
histograms and selected the best. This is now split into two steps:
1. query and compile counts for multiple histograms (currently three)
2. Choose the best one: see `HistogramSelectorComponent`
