+++ src/explorer/Components/NumericHistogramComponent.cs
@@ -6,54 +6,28 @@ namespace Explorer.Components
 
     using Diffix;
     using Explorer.Common;
-    using Explorer.Metrics;
+    using Explorer.Components.ResultTypes;
     using Explorer.Queries;
 
     public class NumericHistogramComponent :
-        ExplorerComponent<NumericHistogramComponent.Result>, PublisherComponent
+        ExplorerComponent<List<HistogramWithCounts>>
     {
         private const long ValuesPerBucketTarget = 20;
         private readonly DConnection conn;
         private readonly ExplorerContext ctx;
-        private readonly ResultProvider<DistinctValuesComponent.Result> distinctValuesProvider;
         private readonly ResultProvider<SimpleStats<double>.Result> statsResultProvider;
 
         public NumericHistogramComponent(
             DConnection conn,
             ExplorerContext ctx,
-            ResultProvider<DistinctValuesComponent.Result> distinctValuesProvider,
             ResultProvider<SimpleStats<double>.Result> statsResultProvider)
         {
             this.conn = conn;
             this.ctx = ctx;
-            this.distinctValuesProvider = distinctValuesProvider;
             this.statsResultProvider = statsResultProvider;
         }
 
-        public async IAsyncEnumerable<ExploreMetric> YieldMetrics()
-        {
-            var distinctValues = await distinctValuesProvider.ResultAsync;
-            if (!distinctValues.IsCategorical)
-            {
-                var result = await ResultAsync;
-
-                if (result.Histogram.Buckets.Count > 0)
-                {
-                    yield return new UntypedMetric("histogram.buckets", result.Histogram.Buckets.Values.Select(b => new
-                    {
-                        BucketSize = b.BucketSize.SnappedSize,
-                        b.LowerBound,
-                        b.Count,
-                        b.CountNoise,
-                    }));
-                    yield return new UntypedMetric("histogram.suppressed_count", result.ValueCounts.SuppressedCount);
-                    yield return new UntypedMetric("histogram.suppressed_ratio", result.ValueCounts.SuppressedCountRatio);
-                    yield return new UntypedMetric("histogram.value_counts", result.ValueCounts);
-                }
-            }
-        }
-
-        protected async override Task<Result> Explore()
+        protected async override Task<List<HistogramWithCounts>> Explore()
         {
             var stats = await statsResultProvider.ResultAsync;
 
@@ -73,31 +47,13 @@ namespace Explorer.Components
                     row => row.BucketSize,
                     (bs, rows) => (BucketSize: new BucketSize(bs), Rows: ValueCounts.Compute(rows)));
 
-            var results = valueCounts.Join(
-                histograms,
-                v => v.BucketSize.SnappedSize,
-                h => h.BucketSize.SnappedSize,
-                (v, h) => new Result(v.Rows, h));
-
-            return results
-                .OrderBy(h => h.BucketSize.SnappedSize)
-                .ThenBy(h => h.ValueCounts.SuppressedCount)
-                .First();
-        }
-
-        public class Result
-        {
-            internal Result(ValueCounts valueCounts, Histogram histogram)
-            {
-                ValueCounts = valueCounts;
-                Histogram = histogram;
-            }
-
-            public BucketSize BucketSize => Histogram.BucketSize;
-
-            public ValueCounts ValueCounts { get; }
-
-            public Histogram Histogram { get; }
+            return valueCounts
+                .Join(
+                    histograms,
+                    v => v.BucketSize.SnappedSize,
+                    h => h.BucketSize.SnappedSize,
+                    (v, h) => new HistogramWithCounts(v.Rows, h))
+                .ToList();
         }
     }
 }
