+++ src/explorer/Components/ResultTypes/HistogramWithCounts.cs
@@ -1,0 +1,19 @@
+namespace Explorer.Components.ResultTypes
+{
+    using Explorer.Common;
+
+    public class HistogramWithCounts
+    {
+        internal HistogramWithCounts(ValueCounts valueCounts, Histogram histogram)
+        {
+            ValueCounts = valueCounts;
+            Histogram = histogram;
+        }
+
+        public BucketSize BucketSize => Histogram.BucketSize;
+
+        public ValueCounts ValueCounts { get; }
+
+        public Histogram Histogram { get; }
+    }
+}
